Fix error in C file 6 ver 4

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	long long int lo;
 	float f;
 
-	printf("Size of a char: %i byte(s)\n", (int)sizeof(c));
-	printf("Size of an int: %i byte(s)\n", (int)sizeof(i));
-	printf("Size of a long: %i byte(s)\n", (int)sizeof(l));
-	printf("Size of a long long: %i byte(s)\n", (int)sizeof(lo));
-	printf("Size of a float: %i byte(s)\n", (int)sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a long: %lu byte(s)\n", (unsigned long)sizeof(l));
+	printf("Size of a long long: %lu byte(s)\n", (unsigned long)sizeof(lo));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	
 	Return (0);
 }
